requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
asgiref==3.8.1
attrs==25.1.0
autocorrect==2.6.1
babel==2.17.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.1
cmake==3.25.2
colorama==0.4.6
comtypes==1.4.9
cv==1.0.0
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2
django-autoslug==1.9.9
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-filter==25.1
django-js-asset==3.1.2
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dlib-bin==19.24.6
docopt==0.6.2
et_xmlfile==2.0.0
Flask==3.1.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.14.0
httpie==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.23.0
nltk==3.9.1
num2words==0.5.14
numpy==2.2.2
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pdfkit==1.0.0
pillow==11.1.0
polib==1.2.0
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
psycopg2==2.9.10
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
PyPDF2==3.0.1
pypiwin32==223
PySocks==1.7.1
pyspellchecker==0.8.2
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.2
pywin32==308
razorpay==1.4.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rjsmin==1.2.4
rpds-py==0.22.3
segno==1.6.6
selenium==4.28.1
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
streamlit==1.41.1
tenacity==9.0.0
textblob==0.19.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
twilio==9.4.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==2.2.3
whitenoise==6.9.0
wsproto==1.2.0
XlsxWriter==3.2.3
yarl==1.18.3
